DataManipulator tests added

+++ src/main/java/com/decentralizeddatabase/reno/crypto/Hasher.java
@@ -14,9 +14,9 @@ public class Hasher {
      * @param blockNumber int
      * @return String (20byte SHA-1 hash)
      */
-    public static String createBlockKey(final String secretKey, final String filename, final int blockNumber){
+    public static String createBlockKey(final String secretKey, final String filename, final long blockNumber){
         final String hashed = Hashing.sha1()
-            .hashString(secretKey + filename + Integer.toString(blockNumber), Charsets.UTF_8)
+            .hashString(secretKey + filename + Long.toString(blockNumber), Charsets.UTF_8)
             .toString();
 
         return hashed;
